requirements.txt
# Core data processing
dbt-duckdb>=1.7.0
duckdb>=0.9.0
python-dateutil>=2.8.0

# Machine Learning & Data Analysis
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.25.0
scipy>=1.10.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.6

# Email validation
email-validator>=2.0.0
phonenumbers>=8.13.0

# Optional: For enhanced logging
rich>=13.0.0 